Extract  into a shared scope to remove duplication
